fixes unsafe comparison

+++ src/process.cpp
@@ -219,7 +219,7 @@ double dist_anchor( const esa_t *C, const char *query, size_t query_length, doub
 	
 	// Abort if we have more homologous nucleotides than just nucleotides. This might
 	// happen with sequences of different lengths.
-	if( homo >= C->len ){
+	if( homo >= (size_t) C->len ){
 		return log(-1.0);
 	}
